Split agent worker stdlib modules
